\name{hicGraph-package}
\alias{hicGraph-package}
\alias{hicGraph}
\docType{package}
\title{
\packageTitle{hicGraph}
}
\description{
\packageDescription{hicGraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hicGraph}
\packageIndices{hicGraph}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hicGraph}

Maintainer: \packageMaintainer{hicGraph}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
